// src/lib/schemas.ts
import { z } from 'zod'

export const ProjectSchema = z.object({
    title: z.string().min(3, { message: 'Title must be at least 3 characters' }),
    description: z.string().min(10, { message: 'Description must be at least 10 characters' }),
    status: z.enum(['active', 'completed', 'archived']),
});

export type Project = z.infer<typeof ProjectSchema>;